Completed authentication system
- Fixed and added many tests.

+++ api/rpc/client/client_account.go
@@ -46,5 +46,5 @@ func GetNextNonceOfAccountUsingRPCClient(address string, client Client) (string,
 	if err != nil {
 		return "", err
 	}
-	return fmt.Sprintf("%d", uint64(acct.Nonce+1)), nil
+	return fmt.Sprintf("%d", acct.Nonce+1), nil
 }
